Move the Order and Reply structs into main
- So we can do go run main.go
